Serde load models skeleton code

+++ src/consts.rs
@@ -1,3 +1,3 @@
 pub const LINE_SEPARATOR:		&str		= "--------------------------------------------------";
 pub const INPUT_IMAGE_SIZE:		(u32, u32)	= (426, 240);
-pub const INTERMEDIATE_LAYERS:	u16			= 126;
+pub const INTERMEDIATE_LAYERS:	usize		= 126;
